adk-capstone/requirements.txt
google-adk[a2a]==1.18.0
google-genai>=0.6.0
fastapi>=0.121.2
uvicorn>=0.30.0
httpx>=0.27.0
pydantic>=2.7
python-dotenv>=1.0

